requirements.txt
flask
gunicorn
numpy
pandas
scikit-learn
pyyaml
six
matplotlib
seaborn
lightgbm
xgboost
pyarrow
-e .